adding code for condensing overview, awaiting cleanup

+++ services/frontend/src/components/overview/overview-task-cell.js
@@ -1,30 +1,64 @@
 import _ from 'lodash';
 import React from 'react';
 import { connect } from 'react-redux';
+import { navigate } from 'gatsby';
 import PropTypes from 'prop-types';
-import { ListItem, ListItemText, Typography } from '@material-ui/core';
-import { getTasksSelectors, updateTask, updateTaskLocal } from '../../store/tasks';
+import {
+    ListItem,
+    ListItemText,
+    Typography,
+    ListItemIcon,
+    ListItemSecondaryAction,
+    IconButton,
+} from '@material-ui/core';
+import EditIcon from '@material-ui/icons/Edit';
+import DeleteIcon from '@material-ui/icons/Delete';
+import { dateUtils } from '@logan/core';
+import { getTasksSelectors, updateTask, updateTaskLocal, setShouldGoToTask } from '../../store/tasks';
 import { getScheduleSelectors } from '../../store/schedule';
 import { getAssignmentsSelectors } from '../../store/assignments';
-import { CourseLabel } from '../shared';
-import globalStyles from '../../globals/global.scss';
+import { CourseLabel, TagsDisplay } from '../shared/displays';
+import { Checkbox } from '../shared/controls';
 import styles from '../tasks/task-cell.module.scss';
+import priorities from '../shared/displays/priority-constants';
+
+const {
+    dayjs,
+    constants: { DB_DATE_FORMAT, DB_DATETIME_FORMAT },
+} = dateUtils;
 
 class OverviewTaskCell extends React.Component {
     constructor(props) {
         super(props);
+
+        this.select = this.select.bind(this);
+        this.deleted = this.deleted.bind(this);
         this.handleChange = this.handleChange.bind(this);
+        this.openRelatedAssignment = this.openRelatedAssignment.bind(this);
+
+        this.shouldShowOverdueLabel = this.shouldShowOverdueLabel.bind(this);
+        this.overdueLabelContent = this.overdueLabelContent.bind(this);
+
         this.state = {
             task: this.props.selectTaskFromStore(this.props.tid),
         };
     }
 
     select() {
-        this.props.onSelect(this.props.tid);
+        if (this.props.onSelect) {
+            this.props.onSelect(this.props.tid);
+        }
     }
 
     deleted() {
-        this.props.onDelete(this.state.task);
+        if (this.props.onDelete) {
+            this.props.onDelete(this.state.task);
+        }
+    }
+
+    openRelatedAssignment() {
+        this.props.setShouldGoToTask(this.props.tid);
+        navigate('/tasks');
     }
 
     componentDidUpdate() {
@@ -40,14 +74,49 @@ class OverviewTaskCell extends React.Component {
             complete: !this.state.task.complete,
         };
 
+        if (changes.complete) {
+            changes.completionDate = dayjs().format(DB_DATETIME_FORMAT);
+        }
+
         this.props.updateTaskLocal({
             id: this.props.tid,
             changes,
         });
 
+        const afterUpdates = _.merge({}, this.state.task, changes);
+
         this.setState({
-            task: _.merge({}, this.state.task, changes),
+            task: afterUpdates,
         });
+
+        this.props.updateTask(afterUpdates);
+    }
+
+    shouldShowOverdueLabel() {
+        if (!this.props.showOverdueLabel) return false;
+        if (!dateUtils.dueDateIsDate(this.state.task.dueDate)) return false;
+
+        const dateValue = dayjs(this.state.task.dueDate, DB_DATE_FORMAT);
+        return dateValue.isBefore(dayjs(), 'day');
+    }
+
+    overdueLabelContent() {
+        const dateValue = dayjs(this.state.task.dueDate, DB_DATE_FORMAT);
+        const days = dayjs().diff(dateValue, 'days');
+
+        if (days === 1) {
+            return 'Due yesterday';
+        } else {
+            return `Due ${days} days ago`;
+        }
+    }
+    determineBoxColor() {
+        const priority = _.get(this.state.task, 'priority');
+        const p = _.defaultTo(
+            _.find(_.values(priorities), ([num]) => num === priority),
+            ['white', 'white']
+        );
+        return p[1];
     }
 
     render() {
@@ -56,35 +125,60 @@ class OverviewTaskCell extends React.Component {
             ? this.props.getCourse(assignment.cid)
             : this.props.getCourse(_.get(this.state.task, 'cid'));
 
-        const needsUpperLabel = course || assignment;
+        const needsUpperLabel = !this.props.subtaskCell && (course || assignment);
         const hasBoth = course && assignment;
 
-        const checkboxStyles = {};
-        if (course && _.get(this.state.task, 'complete')) {
-            checkboxStyles.color = course.color;
-        }
-
         return (
-            <div className={styles.taskCell}>
-                <ListItem>
+            <div className={`list-cell ${styles.taskCell}`}>
+                <ListItem button selected={this.props.selected} onClick={this.select}>
+                    {!this.props.subtaskCell && (
+                        <ListItemIcon>
+                            <Checkbox
+                                color={this.determineBoxColor()}
+                                cid={_.get(this.state.task, 'cid')}
+                                checked={_.get(this.state, 'task.complete', false)}
+                                onChange={this.handleChange}
+                            />
+                        </ListItemIcon>
+                    )}
                     <ListItemText
                         primary={
                             <React.Fragment>
                                 {needsUpperLabel && (
-                                    <div className={globalStyles.cellUpperLabel}>
+                                    <div className="cell-upper-label">
                                         {course && <CourseLabel cid={course.cid} />}
                                         {assignment && (
-                                            <Typography className={globalStyles.assignmentLabel}>
-                                                {(hasBoth && '/') + assignment.title}
+                                            <Typography className="assignment-label">
+                                                {`${hasBoth ? ' / ' : ''}${assignment.title}`}
                                             </Typography>
                                         )}
                                     </div>
                                 )}
                                 <div>{_.get(this.state, 'task.title')}</div>
+                                {_.get(this.state.task, 'tags', []).length > 0 && (
+                                    <TagsDisplay tags={this.state.task.tags} />
+                                )}
+                                {this.shouldShowOverdueLabel() && (
+                                    <Typography variant="body2" color="error">
+                                        {this.overdueLabelContent()}
+                                    </Typography>
+                                )}
                             </React.Fragment>
                         }
                         secondary={_.get(this.state, 'task.description')}
                     />
+                    <ListItemSecondaryAction className="actions">
+                        {this.props.subtaskCell && (
+                            <IconButton edge="end" onClick={this.openRelatedAssignment}>
+                                <EditIcon fontSize="small" />
+                            </IconButton>
+                        )}
+                        {this.props.onDelete ? (
+                            <IconButton edge="end" onClick={this.deleted}>
+                                <DeleteIcon color="error" />
+                            </IconButton>
+                        ) : null}
+                    </ListItemSecondaryAction>
                 </ListItem>
             </div>
         );
@@ -92,7 +186,10 @@ class OverviewTaskCell extends React.Component {
 }
 
 OverviewTaskCell.propTypes = {
+    subtaskCell: PropTypes.bool,
     tid: PropTypes.string,
+    showOverdueLabel: PropTypes.bool,
+    updateTask: PropTypes.func,
     updateTaskLocal: PropTypes.func,
     selectTaskFromStore: PropTypes.func,
     getCourse: PropTypes.func,
@@ -100,6 +197,7 @@ OverviewTaskCell.propTypes = {
     selected: PropTypes.bool,
     onSelect: PropTypes.func,
     onDelete: PropTypes.func,
+    setShouldGoToTask: PropTypes.func,
 };
 
 const mapStateToProps = state => {
@@ -110,6 +208,6 @@ const mapStateToProps = state => {
     };
 };
 
-const mapDispatchToProps = { updateTask, updateTaskLocal };
+const mapDispatchToProps = { updateTask, updateTaskLocal, setShouldGoToTask };
 
 export default connect(mapStateToProps, mapDispatchToProps)(OverviewTaskCell);
